+++ frontend/src/components/Youtube/index.ts
@@ -1,0 +1,4 @@
+export { default as Youtube } from './Youtube';
+export { default as useVideoPlayerContext } from './hooks/useVideoPlayerContext';
+export { VideoPlayerProvider } from './VideoPlayerProvider';
+export { VideoPlayerContext } from './VideoPlayerProvider';
